requirements.txt
Flask
google-auth-oauthlib
google-api-python-client
firebase-admin
pandas